requirements.txt
--index-url https://pypi.python.org/simple/
numpy==1.10.4
scipy==0.17.0
nose==1.3.7
path.py==8.1.2
python-dateutil==2.4.2
joblib==0.9.4
pyprind==2.9.4
mako==1.0.3
-e git://github.com/Theano/Theano.git@9f46b692af32e29b1d0372bc1f260bf542b3f1b9#egg=Theano
-e git://github.com/Lasagne/Lasagne.git@19b9d958507e7300e355d7cd1a6bbff2c7726c07#egg=Lasagne
-e git://github.com/pybox2d/pybox2d.git@8825b16c98a144af53bac9514727c77976a2372c#egg=box2d
hg+http://bitbucket.org/pygame/pygame
